+++ tests/backend_api/test_connector.py
@@ -5,8 +5,10 @@ import pytest
 from requests.exceptions import HTTPError
 
 from mex.common.backend_api.connector import BackendApiConnector
-from mex.common.backend_api.models import ExtractedItemsRequest, PreviewItemsResponse
+from mex.common.backend_api.models import ItemsContainer, PaginatedItemsContainer
 from mex.common.models import (
+    AnyExtractedModel,
+    AnyPreviewModel,
     ExtractedPerson,
     MergedPerson,
     PersonRuleSetRequest,
@@ -45,7 +47,7 @@ def test_post_extracted_items_mocked(
     )
     assert (
         json.loads(mocked_backend.call_args.kwargs["data"])
-        == ExtractedItemsRequest(items=[extracted_person]).model_dump()
+        == ItemsContainer[AnyExtractedModel](items=[extracted_person]).model_dump()
     )
 
 
@@ -202,7 +204,9 @@ def test_fetch_preview_items_mocked(
     mocked_backend: MagicMock,
     preview_person: PreviewPerson,
 ) -> None:
-    preview_response = PreviewItemsResponse(items=[preview_person], total=92)
+    preview_response = PaginatedItemsContainer[AnyPreviewModel](
+        items=[preview_person], total=92
+    )
     mocked_return = preview_response.model_dump()
     mocked_backend.return_value.json.return_value = mocked_return