+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BREAKING: convert write_ndjson to NdjsonSink
 - backend and ndjson sinks log progress only in batches
 - increase timeout and decrease chunk size for backend API sink
+- port backend identity provider implementation from editor/extractors to common
+- allow backend and graph as identity provider setting to simplify setting subclasses,
+  even though graph is not implemented in mex-common
+- BREAKING: make backend api connector response models generic, to keep DRY
 
 ### Deprecated